pyproject.toml
[tool.poetry]
name = "programmer"
version = "0.1.0"
description = ""
authors = ["Belinda Mo <justanexperimentpark@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
openai = "^1.59.9"
requests = "^2.32.3"
dspy = "^2.5.43"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
